core: declare the parameter store once, with slmodemd's signatures
Six files carried their own extern for modem_get_param, four of them saying
int where slmodemd says long.  On x86-32 that is invisible; it is also
undefined behaviour against the harness, which has always defined the long
form, and it truncated the pointer that MDMPRM_DP_ADDR carries -- the one
parameter that is an address rather than a setting.

check64 is syntax-only, so it reported OK while warning about the
truncation in call_of.  It is now silent.

327 groups, 19,805,045 checks, zero failures.

+++ test/harness/runtime.c
@@ -379,7 +379,7 @@ int modem_get_bits(void *m, int nbits, unsigned char *buf, int n)
 int modem_put_bits(void *m, int nbits, const unsigned char *buf, int n)
 { (void)m; (void)nbits; return shim_put_bits(&harness_modem_ours, buf, n); }
 
-int modem_set_param(void *m, unsigned name, int val)
+long modem_set_param(void *m, unsigned name, int val)
 { (void)m; return (int)shim_set_param(&harness_modem_ours, name, val); }
 
 int ref_modem_get_bits(void *m, int nbits, unsigned char *buf, int n)
