+++ test/interop/runtime64.c
@@ -55,6 +55,6 @@ int modem_get_bits(void *m, int c, unsigned char *b, int n)
 { (void)m; (void)c; (void)b; (void)n; return 0; }
 int modem_put_bits(void *m, int c, const unsigned char *b, int n)
 { (void)m; (void)c; (void)b; (void)n; return 0; }
-int modem_set_param(void *m, unsigned p, int v)
+long modem_set_param(void *m, unsigned p, int v)
 { (void)m; (void)p; (void)v; return 0; }
 long modem_get_param(void *m, unsigned p) { (void)m; (void)p; return 0; }
